+++ game/scrabble.py
@@ -92,6 +92,12 @@ class ScrabbleGame:
         elif self.turn > 1: 
             return True;
 
+    def checkIfNextToTile(self, word, location, orientation):
+        if not self.isNextToTile(word, location, orientation) == None:
+            return True;
+        else:
+            raise Exception("Palabra debe continuar con las del tablero!")
+
     def isNextToTile(self, word, location, orientation):
         (x, y) = location
         if self.turn > 1:
@@ -119,11 +125,7 @@ class ScrabbleGame:
             and
             self.checkIfFirstTurn(word, location, orientation)
         ):
-            if not self.isNextToTile(word, location, orientation) == None:
-                word = unidecode(word)
-                return True;
-            else:
-                raise Exception("Palabra debe continuar con las del tablero!")
+            return self.checkIfNextToTile(word, location, orientation)
         elif (
             self.isWordInBoard(word, location, orientation)
             and
@@ -133,20 +135,29 @@ class ScrabbleGame:
             and
             self.checkIfFirstTurn(word, location, orientation)
         ):
-            if not self.isNextToTile(word, location, orientation) == None:
-                return True;
-            else:
-                raise Exception("Palabra debe continuar con las del tablero!")
+            return self.checkIfNextToTile(word, location, orientation)
+
+    def isSpecial(self, letter):
+        if (
+            letter == "ú" 
+            or letter == "é" 
+            or letter == "í"
+            or letter == "ó"
+            or letter == "á"
+        ):
+            letter = unidecode(letter)
+            return letter
+        else:
+            return letter
 
     def putWord(self, word, location, orientation):
         (x, y) = location
         score = []
         if self.validateWord(word, location, orientation):
-            if ('í' in word or 'é' in word or 'ú' in word or 'ó' in word or 'á' in word):
-                word = unidecode(word)
             word = [char for char in word]
             for letter in word:
-                for i in range(len(self.current_player.tiles)):             
+                for i in range(len(self.current_player.tiles)): 
+                    letter = self.isSpecial(letter)
                     if (letter == str(self.board.getCellInBoard(x, y)).lower() and letter == self.current_player.tiles[i].letter.lower()):
                         if orientation == "V" or orientation == "v":
                             score.append(self.board.getCellInBoard(x, y))