+++ run_test.py
@@ -2,6 +2,6 @@ import unittest, pathlib
 
 if __name__ == "__main__":
     loader = unittest.TestLoader()
-    loader.testMethodPrefix = "test_getScore"
+    loader.testMethodPrefix = "test_validateWordMuchasTildes"
     suite = loader.discover(str(pathlib.Path.cwd().joinpath("tests")))
     unittest.TextTestRunner(failfast=True).run(suite)